requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
aliyun-python-sdk-core==2.13.36
aliyun-python-sdk-kms==2.15.0
amqp==5.1.1
anyjson==0.3.3
asgiref==3.5.0
async-timeout==4.0.2
attrs==21.4.0
backports.zoneinfo==0.2.1
base58==2.1.1
billiard==3.6.4.0
bitarray==1.2.2
celery @ https://github.com/celery/celery/tarball/master
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
crcmod==1.7
cryptography==36.0.2
cytoolz==0.11.2
Deprecated==1.2.13
Django==4.0.4
django-ethereum-events==4.2.0
django-solo==2.0.0
djangorestframework==3.13.1
dnspython==2.2.1
eth-abi==2.1.1
eth-account==0.5.7
eth-hash==0.3.2
eth-keyfile==0.5.1
eth-keys==0.3.4
eth-rlp==0.2.1
eth-typing==2.3.0
eth-utils==1.10.0
eventlet==0.33.0
frozenlist==1.3.0
greenlet==1.1.2
hexbytes==0.2.2
idna==3.3
importlib-resources==5.7.1
ipfshttpclient==0.8.0a2
jmespath==0.10.0
jsonschema==4.4.0
kombu==5.2.4
lru-dict==1.1.7
multiaddr==0.0.9
multidict==6.0.2
netaddr==0.8.0
oss2==2.15.0
packaging==21.3
parsimonious==0.8.1
prompt-toolkit==3.0.29
protobuf==3.20.1
pycparser==2.21
pycryptodome==3.14.1
pyparsing==3.0.8
pyrsistent==0.18.1
pytz==2022.1
pywin32==303
redis==4.2.2
requests==2.27.1
rlp==2.0.1
six==1.16.0
sqlparse==0.4.2
toolz==0.11.2
tzdata==2022.1
urllib3==1.26.9
varint==1.0.2
vine==5.0.0
wcwidth==0.2.5
web3==5.29.0
websockets==9.1
wrapt==1.14.0
yarl==1.7.2
zipp==3.8.0
